atmelsysctrl::readAccumulatorStatus: Use 8bits for status instead of 16

+++ mt310s2d/src/micro-controller-io/atmelsysctrl.cpp
@@ -23,7 +23,7 @@ ZeraMcontrollerBase::atmelRM cATMELSysCtrl::readCTRLVersion(QString& answer)
     return readVariableLenText(hwGetCtrlVersion, answer);
 }
 
-ZeraMcontrollerBase::atmelRM cATMELSysCtrl::readAccumulatorStatus(quint16 &stat)
+ZeraMcontrollerBase::atmelRM cATMELSysCtrl::readAccumulatorStatus(quint8 &stat)
 {
     ZeraMcontrollerBase::atmelRM cmdStatus = cmdexecfault;
     quint8 answ[2];
