+++ mt310s2d/testlib/atmelsyscntrltest.h
@@ -7,7 +7,7 @@ class AtmelSysCntrlTest : public cATMELSysCtrl
 {
 public:
     AtmelSysCntrlTest(QString devnode, quint8 adr, quint8 debuglevel);
-    atmelRM readAccumulatorStatus(quint16& stat) override;
+    atmelRM readAccumulatorStatus(quint8& stat) override;
 };
 
 #endif // ATMELSYSCNTRLTEST_H
